Ensure request bodies can be null
