+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/api/JavascriptFirebaseFunctionsHttpsExternals.kt
@@ -13,7 +13,7 @@ internal external class FirebaseFunctionsHttpsRequest {
     val headers: FirebaseFunctionsHttpsParameterMap
     val query: FirebaseFunctionsHttpsParameterMap
     val body: dynamic
-    val rawBody: String
+    val rawBody: String?
 }
 
 internal external class FirebaseFunctionsHttpsParameterMap
