+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/pubsub/PubSub.kt
@@ -4,9 +4,13 @@ import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 
 interface PubSubMessage {
-    val json: Any
+    val json: Any?
 }
 
 inline fun <reified T> PubSubMessage.bodyAsJson(
     jsonSerializer: Json
-): T = jsonSerializer.decodeFromString(string = JSON.stringify(json))
+): T? = json?.let { jsonSerializer.decodeFromString(string = JSON.stringify(json)) }
+
+internal fun FirebaseFunctionsPubSubMessage.toPubSubMessage() = object : PubSubMessage {
+    override val json: Any? = this@toPubSubMessage.json
+}
